# yeonjun/beauty-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Beauty Platform API
  version: 0.1.0
  description: |
    원영님의 뷰티앱(React Native/Expo)과 연동되는 부가 플랫폼 API.
    인증/결제/스토리지/AI 생성/컴플라이언스를 포함.
servers:
  - url: https://api.beauty-platform.example.com
    description: Production
  - url: https://staging-api.beauty-platform.example.com
    description: Staging
security:
  - bearerAuth: []
paths:
...
  /v1/ai/generate-image:
    post:
      summary: AI 이미지 생성 (크레딧 차감 원자적)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateImageRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenerateImageResponse' } } } }
        '402': { description: 크레딧 부족 / 구독 필요 }
        '429': { description: Rate limited }
  /v1/gallery:
    get:
      summary: 내 생성 이미지 목록
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GalleryPage' } } } }
  /v1/gallery/{id}:
    delete:
      summary: 이미지 삭제 (DB 레코드 + R2 객체 원자적)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: No Content }
...
  /v1/push/register:
    post:
      summary: FCM/APNs 토큰 등록
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [platform, token]
              properties:
                platform: { type: string, enum: [ios, android] }
                token: { type: string }
      responses:
        '204': { description: No Content }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    GenerateImageRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string, minLength: 1, maxLength: 500 }
        style:
          type: string
          enum: [natural, glam, vintage, kbeauty]
          default: natural
        referenceImageUrl: { type: string, nullable: true }
    GenerateImageResponse:
      type: object
      properties:
        generationId: { type: string }
        imageUrl: { type: string }
        thumbnailUrl: { type: string }
        creditsRemaining: { type: integer }
        latencyMs: { type: integer }
    GalleryPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              imageUrl: { type: string }
              thumbnailUrl: { type: string }
              prompt: { type: string }
              style: { type: string }
              createdAt: { type: string, format: date-time }
        nextCursor: { type: string, nullable: true }
...