# yeonjun/beauty-platform/server/src/routes/ai.ts
import type { FastifyInstance } from 'fastify';
import { z } from 'zod';
import { randomUUID } from 'node:crypto';
import { requireAuth } from '../middleware/auth.js';
import { supabaseAdmin } from '../lib/supabase.js';
import { generateImage, fetchImageBuffer } from '../lib/ai-image.js';
import { uploadToStorage, deleteFromStorage, AI_BUCKET } from '../lib/storage.js';

const Body = z.object({
  prompt: z.string().min(1).max(500),
  style: z.enum(['natural', 'glam', 'vintage', 'kbeauty']).default('natural'),
  referenceImageUrl: z.string().url().optional(),
});

export default async function aiRoutes(app: FastifyInstance) {
  app.post('/v1/ai/generate-image', { preHandler: requireAuth }, async (req, reply) => {
    const parsed = Body.safeParse(req.body);
    if (!parsed.success) {
      return reply.code(422).send({ code: 'VALIDATION_FAILED', message: parsed.error.message });
    }
    const userId = req.userId!;

    const { data: decResult, error: decErr } = await supabaseAdmin.rpc('decrement_credit', { p_user: userId });
    if (decErr) {
      if (decErr.message.includes('INSUFFICIENT_CREDITS')) {
        return reply.code(402).send({ code: 'INSUFFICIENT_CREDITS', message: 'upgrade to premium or buy credits' });
      }
      return reply.code(500).send({ code: 'DB_ERROR', message: decErr.message });
    }
    const creditsRemaining = (decResult as number) ?? -1;

    let imageUrl: string;
    let latencyMs: number;
    try {
      const res = await generateImage({
        prompt: parsed.data.prompt,
        style: parsed.data.style,
        referenceImageUrl: parsed.data.referenceImageUrl,
      });
      imageUrl = res.imageUrl;
      latencyMs = res.latencyMs;
    } catch (e) {
      app.log.error({ err: e }, 'ideogram failed');
      if (creditsRemaining !== -1) {
        await supabaseAdmin
          .from('profiles')
          .update({ credits_remaining: creditsRemaining + 1 })
          .eq('id', userId);
      }
      return reply.code(502).send({ code: 'UPSTREAM_AI_FAILED', message: 'generation failed' });
    }

    const buf = await fetchImageBuffer(imageUrl);
    const generationId = randomUUID();
    const objectKey = `gen/${userId}/${generationId}.webp`;
    const publicUrl = await uploadToStorage(AI_BUCKET, objectKey, buf, 'image/webp');

    const { error: insErr } = await supabaseAdmin.from('generated_images').insert({
      id: generationId,
      user_id: userId,
      prompt: parsed.data.prompt,
      style: parsed.data.style,
      image_url: publicUrl,
      thumbnail_url: publicUrl,
      object_key: objectKey,
      generation_ms: latencyMs,
    });
    if (insErr) {
      await deleteFromStorage(AI_BUCKET, [objectKey]).catch(() => {});
      return reply.code(500).send({ code: 'DB_ERROR', message: insErr.message });
    }

    return {
      generationId,
      imageUrl: publicUrl,
      thumbnailUrl: publicUrl,
      creditsRemaining,
      latencyMs,
      // AI 기본법 제31조: 생성형 AI 결과물 투명성
      aiDisclosure: {
        isAiGenerated: true,
        model: 'pollinations-ai',
        notice: '이 이미지는 인공지능(AI)에 의해 생성되었습니다. This image was generated by AI.',
      },
    };
  });
}